refactor(creature): extract Proficiency enum to its own file
