+++ cmp-ttrpg-companion/shared/core/src/commonMain/kotlin/com/cyrillrx/rpg/creature/domain/Proficiency.kt
@@ -1,0 +1,14 @@
+package com.cyrillrx.rpg.creature.domain
+
+enum class Proficiency {
+    NONE,
+    PROFICIENT,
+    EXPERT,
+    ;
+
+    fun next(): Proficiency = when (this) {
+        NONE -> PROFICIENT
+        PROFICIENT -> EXPERT
+        EXPERT -> NONE
+    }
+}
